feat: Implement user management features with CRUD operations
- Added UserForm component for creating and editing users.
- Integrated UserTable component to display user list with search and filter functionalities.
- Enhanced userSlice to handle user state, including fetching, adding, updating, and deleting users.
- Updated userThunk to manage asynchronous actions for user operations.
- Improved error handling and loading states in user-related actions.
- Added styling for user form and table components for better UI/UX.
- Implemented role-based filtering and search functionality in UserTable.
- Adjusted API base URL for local development.

+++ client/src/store/api/apiUser.js
@@ -1,14 +1,14 @@
 import api from "./api"
 
 export const addUser = async (userData) => {
-  const response = await api.post("/users", userData);
+  const response = await api.post("/auth/register", userData);
   return response;
 }
 
 export const getUser = async () => {
   const response = await api.get("/auth/users");
   console.log("getUser", response);
-  
+
   return response;
 }
 
@@ -18,6 +18,11 @@ export const getUserById = async (id) => {
 }
 
 export const deleteUserById = async (id) => {
-  const response = await api.delete(`/users/${id}`);
+  const response = await api.delete(`/auth/users/${id}`);
+  return response;
+}
+
+export const updateUserById = async (id, userData) => {
+  const response = await api.put(`/auth/users/${id}`, userData);
   return response;
 }
